Add registry hooks for vpnservice
We now send out PRECOMMIT_ and AFTER_ events for CREATE, UPDATE and
DELETE via the default neutron registry. This allows custom drivers or
external software to react on vpnservice events. For custom drivers this
is currently only possible on IPSecSiteConnections. For other objects
the driver gets a method call, but doesn't have the change to do any
validation on its own.