Add custom asr1k-vpnaas extension
For our asr1k driver we have extra supported extensions
(vpn-asr1k-int-tunnel-ips for internal tunnel ips and
vpn-asr1k-nat-address for tunnel nat addresses). As we don't want to
override this in the default driver and we also don't know of another
way to override the supported_extension_aliases we now have our own
plugin.

Originally we planned on keeping the plugin in our asr1k driver repo,
but this did not work, as it couldn't find the all vpn/vpnaas
extensions. The reason for this is that neutron loads extensions for all
service plugins by looking at the base module of each service plugin and
then using this to search them. If we keep our plugin in neutron-vpnaas,
everything can be found. The other alternative would be to set
api_extensions_path and set this to neutron-vpnaas. Details can be found
in neutron.api.extensions.py get_extensions_path().