Add registry hooks for some VPNaaS objects
We now send out PRECOMMIT_ and AFTER_ events for CREATE, UPDATE and
DELETE via the default neutron registry for VPNService,
IPSecSiteConnection and Endpoint Groups. This allows custom drivers or
external software to react on these events. For custom drivers this was
only possible for IPSecSiteConnections. For other objects the driver
gets a method call, but only after everything has already been committed
to the DB, so drivers didn't have any chance to do their own validation.
In the case of IPsecSiteConnection we also add an extra BEFORE_DELETE
hook to have the chance to fetch some related objects before the
connection is actually deleted (i.e. objects that would be gone in
DELETE_PRECOMMIT, but still exist in BEFORE_DELETE due to a foreign key
constraint).

Some of these objects don't support being updated while in use, but the
hooks have been added anyway, just to be prepared for the eventuality
that this changes at some point.